Introduce an event/hook system based on pluggy
This commit introduces a shared hook system for all projects in
the pubtools namespace.

Why do we need this: in summary, we need a way for one project to
be able to make something happen in another project (or in Pub)
without a hard dependency between the two (and without circular
dependencies).

Current use-cases for this include:

- When a Pulp repo is published in pubtools-pulp, we should flush
  CDN cache. But the CDN cache flushing code ideally doesn't live
  in pubtools-pulp since it's not specific to Pulp.

- Translate certain events into messages on a message bus. For example,
  pubtools-quay can invoke a 'quay_removed_repositories' hook when repos
  were removed. If we want to, we can have Pub attach to that hook and
  send UMB messages when invoked, without needing pubtools-quay to learn
  anything about UMB.

- Permission checks: we haven't had a good solution for doing additional
  permission checks in the middle of a decoupled task. The idea
  previously has been that all permission checks must take place before
  the task's entry point is invoked to avoid the task having any hard
  dependency on Pub.

  The hook system makes a new approach possible: the task can emit an
  event communicating what it's about to do, e.g. 'clear_repo_start'.
  If Pub wants to require that the current user needs certain
  permissions for that to happen, it can connect to the event, check the
  permission and raise an exception if policy says it should not be
  allowed.

- Testing hooks: when one project is about to create an object such
  as a client for some remote system, it can invoke a hook. The hook
  can be overridden in the test suite of another project to return
  a mock instead, if desired.

  While technically already possible with mock.patch, that strongly
  relies on implementation details and can be broken simply by renaming
  an internal module; having hooks explicitly designed to enable this
  kind of testing is more maintainable.

Why pluggy specifically: because this is used by both pytest and tox,
which means we're already using it everywhere (even on py2.6
environments). Therefore it seems to be a safe choice.

About docs: the idea is that all hooks in all pubtools projects can be
automatically discovered and included in the docs here. However, I can't
complete this yet since we need to make the first release of pubtools
with this functionality before anyone can define their hooks. The doc
generation script will need some slight updates after that happens.