hooks: add eager loading of entry points to ensure initialization
The recommended approach to registering hookimpls and hookspecs is to
register them when your module is imported.

Problem: there can be modules which want to declare hooks but
wouldn't be imported during the natural course of events while running a
task. Hence, hooks could be missing.

The mkhooks script in this repo already ran into this problem and it
attempted to solve it by eagerly importing all console_scripts. That
works for projects like pubtools-quay, but is insufficient in general
because there may be "hook-only" libraries which have no
console_scripts.

For example, consider a hypothetical pubtools-umb library with desired
behavior: if it's installed, and certain env vars are set, it will
attach to various hooks and send messages on a message bus when those
are activated.

Such a project has no need of any console_script, and we also don't want
to require each project to "import pubtools.umb" just to activate it.
So, there is no obvious way to arrange for the relevant module to be
imported.

This commit provides a solution: a hook-only library such as
pubtools-umb can declare an entry point in pubtools.hooks, and
task_context will ensure any such entry points are fully resolved before
proceeding.