Drop black from github actions CI
"tox -e static" previously ran the black autoformatter using a version
managed by pip-compile.

Problem: pre-commit also runs black, but this uses the version defined
in the pre-commit config. If there are any differences in the black code
style between the two versions of the tool, they will not both be able
to pass.

This recently blocked dependency updates for this repo because a new
release of black is available which has some incompatibilities with the
version used by pre-commit.

Rather than trying to keep both versions in sync, it makes more sense to
apply the tool in just one place; so let's drop it from tox & github
actions and rely only on pre-commit.ci to both run the tool and manage
the version of the tool.