Bump python version in CI to 3.10
Update the python version used for various CI workflows.
Includes a fresh run of "pip-compile" on the updated python.

This does not reflect an intended change in the python versions
supported by this library. The motivation is that some of the CI
configurations install pubtools-quay to help exercise the hooks
mechanism or to generate docs, and that project no longer works on
Python versions older than 3.10 because it started to use
PEP604-style unions in type hints.