Fix issues in otel tests causing failures after upgrading deps
These tests tried to ensure that a TracingWrapper existed with tracing
enabled by doing:

    monkeypatch.setenv("OTEL_TRACING", "true")
    monkeypatch.setenv("OTEL_SERVICE_NAME", "local-test2")
    tw = get_trace_wrapper()

However, that's not guaranteed to work because get_trace_wrapper() might
have already been called prior to the test, while OTEL_TRACING was NOT
set to "true", initializing the global singleton to an object with
tracing disabled.

The most straightforward way to fix this would initially seem to be:
make the tests not use the singleton but instead create a fresh
TracingWrapper for each one. Unfortunately that doesn't work because
opentelemetry library itself has some singletons which can't be adjusted
after initialization, forcing us to only use a singleton as well.

So, provide a private _reset method just for these tests which can
ensure that the TracingWrapper definitely has tracing enabled, even if
it was previously initialized with tracing disabled.