Make open-telemetry a soft dependency [RHELDST-27627]
Several client tools (e.g. rhsm-tools) make use of various parts of
pubtools for interaction with pulp. This has caused issues with
rcm-dev, as we have specifically avoided making open-telemetry
packages available for CLI clients.

This change introduces a new TracingWrapper class that acts as a pass
through for when opentelemetry isn't installed.